refactor(ios): extract merch-artifact decoding from MobileChatContentParser (#14349)
## Summary

Repo-quality loop, iteration 2 (no Linear issue — ad-hoc loop). Separates merch-artifact decoding from the chat content parser. `MobileChatContentParser.swift` **824 → 659 lines**.

- Merch decoding → new `Core/MobileChatMerchParsing.swift` as an `extension MobileChatContentParser` (171 lines). A distinct responsibility — turning tool-result JSON into merch cards — not core prose/tool-call parsing.
- Access: the 5 symbols crossing the file boundary go `private` → internal (4 called from core, plus `firstCapture` called from merch). Merch-internal helpers stay `private`. All bodies verbatim.
- `project.pbxproj` wired (`plutil -lint` clean); brace balance verified.

Note: the residual parser core (659) is a tightly-coupled unit (segmentation ↔ tool-result parsing share a dense private call graph); further splitting has poor risk/reward, so this is the clean seam.

## Verification

- `pnpm run ios:lint` → **clean ✓**
- `pnpm run ios:test` → **TEST SUCCEEDED** (full pass, incl. `MobileChatContentParserTests`).

## Risks

- Low: same-type extension move, 5 access widenings, no logic changes.
